chore: Fix permissions for LOG_FILE in logger function
The code changes in the `logger` function fix the permissions for the `LOG_FILE` if it is not writable. If the `LOG_FILE` does not exist, it will be created and the permissions will be set to 644. This ensures that the `LOG_FILE` is set and writable, preventing any errors related to writing logs.