feat: Add image file format to image download process
The commit adds the `image_file_format` variable to the image download process. This variable is extracted from the JSON file and used to determine the format of the downloaded image. It ensures that the correct file is being targeted is used when performing integrity checks on the downloaded images.